fix: serialize parse_error to frontend (skip_deserializing, not skip)
Replace #[serde(skip)] with #[serde(default)] on parse_error field so the
field is included in serialization to the frontend and in specta TypeScript
bindings. Also change ParseError.line_number from usize to u32 to satisfy
specta's BigInt restriction. Cast idx+1 to u32 in the parser accordingly.

+++ src-tauri/src/model.rs
@@ -43,13 +43,14 @@ pub struct ConversationEvent {
     #[serde(default)]
     pub tool_use_result: Option<serde_json::Value>,
     /// Set by the parser when the raw line could not be decoded as JSON.
-    #[serde(skip)]
+    /// Never present in source `.jsonl`; synthesized by the parser only.
+    #[serde(default)]
     pub parse_error: Option<ParseError>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
 pub struct ParseError {
-    pub line_number: usize,
+    pub line_number: u32,
     pub raw: String,
     pub message: String,
 }
